Added calculate price method to each of my subclasses

+++ AOC2_Week1/AOC2_Week1/macbookPro.h
@@ -11,6 +11,6 @@
 @interface macbookPro : appleProductsBaseClass
 
 @property BOOL retenia;
-@property float screenSizeInInches;
+@property int screenSizeInInches;
 
 @end
